Rename files according to class and export name
