+++ kratos/utilities/parallel_distance_calculator.h
